net: phy: marvell10g: force XFI/SGMII mode
The N366 strapped rate matching mode relies on the MAC to throttle
egress traffic when the serdes rate (10Gbps) is higher than the Ethernet
wire rate. See section 3.4.1 (Rate Matching) in rev C of the datasheet.

We currently don't do anything in the MAC to throttle egress. This might
be causing high packet loss and packets corruption when Ethernet rate is
set to 1G, and egress rate is close to 1G.