net: sfp: workaround missing Tx disable pull-up
When Tx disable pull-up is missing the SFP module might not sense the
transition from disable to enable. The signal just stays low.

As a workaround assert Tx disable on probe.

This only works when the SFP is plugged in when the sfp module probe.
Hot plug of SFP module might not work.

Signed-off-by: Baruch Siach <baruch@tkos.co.il>