net: mvpp2: tai: add extts support
This commit add support for capturing
a timestamp in which the PTP_PULSE pin,
received a signal.

This feature is needed in order to synchronize
multiple clocks in the same board, using
tools like ts2phc from the linuxptp project.

On the Armada 8040, this is the only way to
do so as a result of multiple erattas with
the PTP_PULSE_IN interface that was designed
to synchronize the TAI on an external PPS
signal (the errattas are FE-6856276, FE-7382160
from document MV-S501388-00).

This patch introduces a pinctrl
configuration "extts" that will be selected
once the user had enabled extts, and then
will be returned back to the "default"
pinctrl config once it has been disabled.

This pinctrl mess is needed due to the fact that
there is no way for us to distinguish between
an external trigger (e.g. from the PTP_PULSE_IN
pin) or an internal one, triggered by the registers.

An example should be:

&cp1_ethernet {
	pinctrl-names = "default", "extts";
	pinctrl-0 = <&cp1_mpp6_gpio>;
	pinctrl-1 = <&cp1_mpp6_ptp>;
	status = "okay";
};

This feature has been tested on an Aramda
8040 based board, with linuxptp 3.1.1's ts2phc.

Signed-off-by: Shmuel Hazan <shmuel.h@siklu.com>