DEV-6442: Add initialization order to the falcon phy and modems
The way to receive this order is to define the dpll as hw clock (in falcon.dts, and clock registration in ptp_clockmatrix.c dpll driver init function), and rely on the dts clock dependencies (declare both phy and pcie as depended on this clock).

This works well for the phy but the pcies (modem) need to enable two clocks ( CP110_GATE_PCIE_X1_0 & CP110_GATE_PCIE_X4) that were formerly disabled by the kernel, and that causes kernel stuck. (in the former version, the pcie driver init was earlier and before the kernel disabled those clocks). Also reset pcie through gpio after the disable+enable didn't help. So "DEV-6467: Improve DEV-6442-Add initialization order to the falcon phy and modems" was opened to further investigate and handle this problem, and meanwhile a patch was entered in cp110-system-controller.c:cp110_gate_disable to never disable those two clocks.